flis/trunk/requirements.txt
Flask==0.9
Flask-Script==0.3.3
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
blinker==1.2
distribute==0.6.24
flatland==dev
htables==0.4b5
simplejson==2.6.1
path.py==2.3
psycopg2==2.4.5
-e git+git://github.com/eaudeweb/Flask-HTables.git@master#egg=Flask_HTables-dev
-e git+git://github.com/eaudeweb/flatkit.git@master#egg=flatkit-dev
supervisor==3.0a12
Werkzeug==0.8.3
